requirements.txt
altgraph==0.17.4
astropy==7.1.0
astropy-iers-data==0.2025.7.7.0.39.39
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.6
contourpy==1.3.2
corner==2.2.3
customtkinter==5.2.2
cycler==0.12.1
Cython==3.1.2
darkdetect==0.8.0
decorator==5.2.1
emcee==3.1.6
flexcache==0.3
flexparser==0.4
fonttools==4.58.5
GPy==1.13.2
idna==2.10
Jinja2==3.0.1
jplephem==2.23
kiwisolver==1.4.8
loguru==0.7.3
MarkupSafe==2.0.1
matplotlib==3.10.3
nestle==0.2.0
numdifftools==0.9.41
numpy==1.26.4
packaging==25.0
paramz==0.9.6
pefile==2023.2.7
pillow==11.3.0
Pint==0.24.4
pint-pulsar==1.1.3
platformdirs==4.3.8
pyerfa==2.0.1.5
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.25.1
scipy==1.12.0
setuptools==80.9.0
six==1.17.0
tkinterdnd2==0.4.3
typing_extensions==4.14.1
uncertainties==3.2.3
urllib3==1.26.6
win32_setctime==1.2.0
